+++ CHANGELOG.rst
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 
 Inspired by [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+[pending]
+=========
+Added
+-----
+* `astwro.utils.TempDir` is noe `os.PathLike`.
+
 [0.7.3]
-============
+=======
 Added
 -----
 * `astwro.coord`: `box` and `central` convenience methods for finding ceneter and box boundaty of coors array